% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onedriver_excel.R
\name{onedriver_excel}
\alias{onedriver_excel}
\title{Onedrive Excel}
\usage{
onedriver_excel(sharedURL, file_name, save2wd = FALSE)
}
\arguments{
\item{sharedURL}{shared link from OneDrive excel file}

\item{file_name}{Local file name}

\item{save2wd}{save the file to their working directory?}
}
\description{
Onedrive Excel
}
\examples{
url <- "https://1drv.ms/x/s!AvKSrDD28Jp5q-MDRzpOYGOMSTtmqg?e=kMpWaa"
onedriver_excel(url, file_name = "example.xlsx")

}
