% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onedriver_excel.R
\name{onedriver_excel}
\alias{onedriver_excel}
\title{Onedrive Excel}
\usage{
onedriver_excel(shared_url, file_name, save2wd = FALSE)
}
\arguments{
\item{shared_url}{Excel file shared link on onedrive}

\item{file_name}{Local file name}

\item{save2wd}{save the file to their working directory?}
}
\description{
Onedrive Excel
}
\examples{
url <- "https://ufscbr-my.sharepoint.com/:x:/g/personal/joao_tolentino_ufsc_br/EZaB18_ZxhVGrwsuVsv83c4BsKk3r3jp3r-KEZmWITscGQ?e=aKMKiD"
onedriver_excel(url, file_name = "exemplo.xlsx")

}
